strucpp_abi: move from core/include/ to core/src/lib/
core/include/ was a directory I introduced when I added the ABI
mirror; the existing project convention is core/src/lib/ for runtime
library headers (where iec_types.h already lives). Moving the file
there matches the layout, removes the extra top-level dir, and lets
the .cpp callers use the same `../lib/foo.h` relative-include pattern
they already use for iec_types.h.

Updates:
- git mv core/include/strucpp_abi.hpp → core/src/lib/strucpp_abi.hpp
- image_tables.cpp + plc_state_manager.cpp: include via "../lib/strucpp_abi.hpp"
- CMakeLists.txt: drop core/include from include_directories
- Empty core/include/ directory removed
