+++ core/src/lib/strucpp_abi.hpp
